requirements.txt
torch>=2.0.0
torchvision>=0.19.0
pyyaml
decord
fastapi==0.112.2
pycryptodomex==3.20.0
uvicorn
requests
aiohttp
colorlog
psutil
pydantic==2.8.2
pydantic_core==2.20.1
transformers>=4.30.0
bitsandbytes>=0.39.0
accelerate>=0.22.0
opencv-python>=4.7.0
Pillow>=9.0.0
